Move genTreeOps to its own header so simd.h can use it in Evaluate*Simd
